After several copy/paste action, Vector3 is out and PVector is in. Examples are updated to work accordingly. Some changes are still needed but the core is there.

+++ branches/3d/src/net/nexttext/property/NumberProperty.java
@@ -30,24 +30,24 @@ package net.nexttext.property;
 /* $Id$ */
 public class NumberProperty extends Property {
 
-    double original;
-    double value;
+    float original;
+    float value;
   
-    public NumberProperty(double value) {
+    public NumberProperty(float value) {
         original = value;
         this.value = value;      
     }
 
-    public double getOriginal() { return original; }
+    public float getOriginal() { return original; }
 
-    public double get() { return value; }
+    public float get() { return value; }
 
-    public void set(double value) { 
+    public void set(float value) { 
        this.value = value;
        firePropertyChangeEvent();
     }
 
-    public void add(double value) {
+    public void add(float value) {
         this.value += value;
         firePropertyChangeEvent();
     }
@@ -64,8 +64,8 @@ public class NumberProperty extends Property {
     // Accessors in case it's easier to treat a number property as a long.
 
     public NumberProperty(long value) {
-        original = (double) value;
-        this.value = (double) value;     
+        original = (float) value;
+        this.value = (float) value;     
     }
 
     public long getOriginalLong() { return (long) original; }
@@ -74,7 +74,7 @@ public class NumberProperty extends Property {
     }
 
     public void set(long value) { 
-        this.value = (double) value;
+        this.value = (float) value;
         firePropertyChangeEvent();
     }
